amend: move joinwhenarray to utils
