--- mainapp/src/Component/joinWhenArray.tsx
@@ -1,3 +1,0 @@
-export function joinWhenArray(allowedSources: string | string[]): string {
-    return Array.isArray(allowedSources) ? allowedSources.join(";") : allowedSources;
-}
